add block extra
